rejection/src/requirements.txt
matplotlib==3.8.2
numpy==1.25.0
pandas==2.0.2
Pillow==9.2.0
Pillow==11.1.0